requirements.txt
streamlit
deep-translator
easyocr
pdf2docx
pdf2image
Pillow
torch
